pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ssl-rag-research"
version = "1.0.0"
description = "SSL corpus (PDF + web) embedding pipeline and RAG V1 research codebase"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
    "rank_bm25>=0.2.2",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "pydantic>=2.0",
    "pypdf>=3.0",
    "tiktoken>=0.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest"]
